// src/common/types/withModel.ts
export interface IWithModelResponse {
    pageable: {
        pageSize: number; // required, количество элементов на странице
        pageIndex: number; // required, номер страницы
        totalPages: number; // required, кол-во страниц
        totalElements: number; // required, общее кол-во элементов
    };
    content: IWithModelItem[];
}